requirements.txt
asgiref==3.3.1
astroid==2.4.2
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.3
django-ajax==0.3.0
django-cors-headers==3.7.0
django-crispy==1.8.1
django-debug-toolbar==3.2
django-filter==2.4.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.0.4
isort==5.7.0
lazy-object-proxy==1.4.3
Markdown==3.3.4
mccabe==0.6.1
Pillow==8.0.1
psycopg2==2.8.6
Pygments==2.8.1
pylint==2.6.0
python-decouple==3.4
python-dotenv==0.15.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
wrapt==1.12.1
